Edited the field format for ammount in migrations, rounded up the data outputted to sidebar. Included the artisan function for reward (to be included in cron).

+++ app/database/migrations/2014_09_18_145146_user_tables_create.php
@@ -90,7 +90,7 @@ class UserTablesCreate extends Migration {
 		Schema::create('users_transaction', function($table)
         {
             $table->increments('id');
-            $table->integer('ammount');
+            $table->float('ammount');
             $table->timestamp('date');
             $table->string('payment_system')->nullable();
             $table->integer('confirmed')->nullable();
